+++ lib/src/services/web_socket_service.dart
@@ -5,6 +5,7 @@ import 'dart:typed_data';
 import 'package:logging/logging.dart';
 import 'package:openim_sdk/openim_sdk.dart';
 import 'package:openim_sdk/src/enums/web_socket_status.dart';
+import 'package:universal_platform/universal_platform.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 
 import '../models/web_socket_codec.dart';
@@ -88,8 +89,9 @@ class WebSocketService {
   }) async {
     _userID = userID;
     _token = token;
-    _compression = compression;
-    _codec = WebSocketCodecs(enableCompression: compression);
+    // dart:io 的 gzip 在 Web 平台不可用，自动禁用压缩
+    _compression = UniversalPlatform.isWeb ? false : compression;
+    _codec = WebSocketCodecs(enableCompression: _compression);
     _userDisconnected = false;
     _isReconnecting = false;
     _reconnectAttempts = 0;